+++ server/src/functions.ts
@@ -55,6 +55,7 @@ export const hashParms = (uri: string): any => {
 }
 
 export const isAbap = (uri: string) => !!uri.match(/\.abap$/i)
+export const isCdsView = (uri: string) => !!uri.match(/\.ddls.asddls$/i)
 
 interface RunningState<T> {
   current: Promise<T>